refactor(cli): isolate storage cleanup
