+++ src/bin/dolos/main.rs
@@ -9,6 +9,8 @@ mod eval;
 mod feedback;
 mod serve;
 mod snapshot;
+#[cfg(feature = "utils")]
+mod storage_cleanup;
 mod sync;
 
 #[cfg(feature = "utils")]
